introduction on settings to organise left side
